+++ app/src/main/java/com/godwpfh/myapplication/adapter/ProfileViewPagerAdpater.kt
@@ -1,14 +1,15 @@
-package com.godwpfh.myapplication.home.adapter
+package com.godwpfh.myapplication.adapter
 
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
 
-class TabProfileAdpater(fragment: Fragment) :
-    FragmentStateAdapter(fragment) {
+class ProfileViewPagerAdpater(fragmentActivity: FragmentActivity)
+    :FragmentStateAdapter(fragmentActivity){
 
     private val fragments= mutableListOf<Fragment>()
 
-    override fun getItemCount(): Int= fragments.size
+    override fun getItemCount(): Int = fragments.size
 
     override fun createFragment(position: Int): Fragment = fragments[position]
