+++ app/src/main/java/com/godwpfh/myapplication/adapter/TabProfileAdpater.kt
@@ -1,15 +1,14 @@
-package com.godwpfh.myapplication.home.adapter
+package com.godwpfh.myapplication.adapter
 
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
 
-class ProfileViewPagerAdpater(fragmentActivity: FragmentActivity)
-    :FragmentStateAdapter(fragmentActivity){
+class TabProfileAdpater(fragment: Fragment) :
+    FragmentStateAdapter(fragment) {
 
     private val fragments= mutableListOf<Fragment>()
 
-    override fun getItemCount(): Int = fragments.size
+    override fun getItemCount(): Int= fragments.size
 
     override fun createFragment(position: Int): Fragment = fragments[position]
