+++ app/src/main/java/com/godwpfh/myapplication/data/remote/SoptService.kt
@@ -1,5 +1,9 @@
-package com.godwpfh.myapplication
+package com.godwpfh.myapplication.data.remote
 
+import com.godwpfh.myapplication.data.remote.request.RequestSignIn
+import com.godwpfh.myapplication.data.remote.request.RequestSignUp
+import com.godwpfh.myapplication.data.remote.response.ResponseSignIn
+import com.godwpfh.myapplication.data.remote.response.ResponseSignUp
 import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.POST
