+++ app/src/main/java/com/godwpfh/myapplication/data/remote/response/ResponseSignUp.kt
@@ -1,4 +1,4 @@
-package com.godwpfh.myapplication
+package com.godwpfh.myapplication.data.remote.response
 
 data class ResponseSignUp(
     val status: Int,
