+++ app/src/main/java/com/godwpfh/myapplication/ui/home/TabHomeFollowerFragment.kt
@@ -1,13 +1,11 @@
-package com.godwpfh.myapplication.home
+package com.godwpfh.myapplication.ui.home
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.godwpfh.myapplication.R
 import com.godwpfh.myapplication.databinding.FragmentTabHomeFollowerBinding
-import com.godwpfh.myapplication.home.adapter.TabProfileAdpater
 
 class TabHomeFollowerFragment : Fragment() {
     private var _binding: FragmentTabHomeFollowerBinding?=null
