+++ app/src/main/res/layout/activity_sign_in.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".auth.SignInActivity">
+    tools:context=".ui.auth.SignInActivity">
 
     <ImageView
         android:id="@+id/imageview_github"
