+++ app/src/main/res/layout/fragment_follow.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".home.FollowFragment">
+    tools:context=".ui.home.FollowFragment">
 
 
     <androidx.recyclerview.widget.RecyclerView
