+++ app/src/main/res/layout/fragment_profile.xml
@@ -1,10 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <data>
+        <import type="android.view.View"/>
+        <variable
+            name="viewModel"
+            type="com.godwpfh.myapplication.ui.home.HomeViewModel" />
+    </data>
+
+<androidx.constraintlayout.widget.ConstraintLayout
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".home.ProfileFragment">
+    tools:context=".ui.home.ProfileFragment">
 
 
     <de.hdodenhof.circleimageview.CircleImageView
@@ -28,7 +38,7 @@
         android:layout_marginStart="5dp"
         android:layout_marginTop="10dp"
         android:fontFamily="@font/noto_sans_regular"
-        android:text="@string/user_name"
+        android:text="@{viewModel.name}"
         android:textSize="15sp"
         android:textStyle="bold"
         app:layout_constraintLeft_toLeftOf="parent"
@@ -72,3 +82,4 @@
         app:layout_constraintTop_toBottomOf="@id/home_repos_btn" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
