+++ app/src/main/res/layout/fragment_repos.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".home.ReposFragment">
+    tools:context=".ui.home.ReposFragment">
 
 
     <androidx.recyclerview.widget.RecyclerView
