+++ app/src/main/res/layout/fragment_tab_home_follower.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/sopt_gray"
-    tools:context=".home.TabHomeFollowerFragment">
+    tools:context=".ui.home.TabHomeFollowerFragment">
 
     <TextView
         android:layout_width="wrap_content"
